As part of a desire to reduce code duplication, I am trying a number of ideas. One is to implement everything using void pointers in C, which makes things difficult to read but generic. Another is to implement everything in C++, which does provide facilities for generic programming (templates). These files are the beginnings of an attempt to implement photon queues, whose performance can be compared directly to the C program channels.
So far, the C++ code is slower by about a factor of 100 in system time and 10 in wall time. I played around with increasing the file buffer size a bit, but have not had all that much luck so far. Even so, keep pushing at this, as it does provide a simpler interface for advanced tasks and the performance issues can probably be mitigated.